// dotcom-rendering/src/model/zod/commercial.ts
// Generated by ts-to-zod
import type { output } from 'zod';
import { z } from 'zod';
import { brandingSchema } from './branding';

const adTargetParamSchema = z.object({
	name: z.string(),
	value: z.union([z.string(), z.array(z.string())]),
});

const editionCommercialPropertiesSchema = z.object({
	adTargeting: z.array(adTargetParamSchema),
	branding: brandingSchema.optional(),
});

export type ZCommercialProperties = output<typeof commercialPropertiesSchema>;
export const commercialPropertiesSchema = z.object({
	UK: editionCommercialPropertiesSchema,
	US: editionCommercialPropertiesSchema,
	AU: editionCommercialPropertiesSchema,
	INT: editionCommercialPropertiesSchema,
});