// dotcom-rendering/src/model/zod/elements.ts
const audioBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.AudioBlockElement'),
	elementId: string(),
});

const blockquoteBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.BlockquoteBlockElement'),
	elementId: string(),
	html: string(),
	quoted: boolean().optional(),
});

const captionBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.CaptionBlockElement'),
	elementId: string(),
	captionText: string().optional(),
	padCaption: boolean().optional(),
	credit: string().optional(),
	displayCredit: boolean().optional(),
	shouldLimitWidth: boolean().optional(),
	isOverlaid: boolean().optional(),
});

const codeBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.CodeBlockElement'),
	elementId: string(),
	html: string(),
	isMandatory: boolean(),
	language: string().optional(),
});

const contentAtomBlockElementSchema = object({
	_type: literal(
		'model.dotcomrendering.pageElements.ContentAtomBlockElement',
	),
	elementId: string(),
	atomId: string(),
});

const dividerBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.DividerBlockElement'),
	size: union([literal('full'), literal('partial')]).optional(),
	spaceAbove: union([literal('tight'), literal('loose')]).optional(),
});

const genericAtomBlockElementSchema = object({
	_type: literal(
		'model.dotcomrendering.pageElements.GenericAtomBlockElement',
	),
	url: string(),
	placeholderUrl: string().optional(),
	id: string().optional(),
	html: string().optional(),
	css: string().optional(),
	js: string().optional(),
	elementId: string(),
});

const highlightBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.HighlightBlockElement'),
	elementId: string(),
	html: string(),
});

const itemLinkBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.ItemLinkBlockElement'),
	elementId: string(),
	html: string(),
});

const numberedTitleBlockElementSchema = object({
	_type: literal(
		'model.dotcomrendering.pageElements.NumberedTitleBlockElement',
	),
	elementId: string(),
	position: number(),
	html: string(),
	format: capiFormatSchema,
});

const pullquoteBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.PullquoteBlockElement'),
	elementId: string(),
	html: string().optional(),
	role: string(),
	attribution: string().optional(),
	isThirdPartyTracking: boolean().optional(),
});
// ...
const audioAtomBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.AudioAtomBlockElement'),
	elementId: string(),
	id: string(),
	kicker: string(),
	title: string().optional(),
	trackUrl: string(),
	duration: number(),
	coverUrl: string(),
	role: roleTypeSchema.optional(),
});

const chartAtomBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.ChartAtomBlockElement'),
	elementId: string(),
	id: string(),
	url: string(),
	html: string(),
	css: string().optional(),
	js: string().optional(),
	role: roleTypeSchema.optional(),
	placeholderUrl: string().optional(),
});

const quizAtomBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.QuizAtomBlockElement'),
	elementId: string(),
	quizType: enums(['personality', 'knowledge']),
	id: string(),
	questions: array(questionTypeSchema),
	resultBuckets: array(resultBucketsTypeSchema),
	resultGroups: array(
		object({
			id: string(),
			title: string(),
			shareText: string(),
			minScore: number(),
		}),
	),
});

const commentBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.CommentBlockElement'),
	elementId: string(),
	body: string(),
	avatarURL: string(),
	profileURL: string(),
	profileName: string(),
	permalink: string(),
	dateTime: string(),
	role: roleTypeSchema.optional(),
});

const disclaimerBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.DisclaimerBlockElement'),
	elementId: string(),
	html: string(),
	role: roleTypeSchema.optional(),
});

const documentBlockElementSchema = thirdPartyEmbeddedContentSchema.extend({
	_type: literal('model.dotcomrendering.pageElements.DocumentBlockElement'),
	elementId: string(),
	embedUrl: string(),
	height: number(),
	width: number(),
	title: string().optional(),
	role: roleTypeSchema.optional(),
});

const embedBlockElementSchema = thirdPartyEmbeddedContentSchema.extend({
	_type: literal('model.dotcomrendering.pageElements.EmbedBlockElement'),
	elementId: string(),
	safe: boolean().optional(),
	role: roleTypeSchema.optional(),
	alt: string().optional(),
	height: number().optional(),
	width: number().optional(),
	html: string(),
	isMandatory: boolean(),
	caption: string().optional(),
});

const explainerAtomBlockElementSchema = object({
	_type: literal(
		'model.dotcomrendering.pageElements.ExplainerAtomBlockElement',
	),
	elementId: string(),
	id: string(),
	title: string(),
	body: string(),
	role: roleTypeSchema.optional(),
});

const guideAtomBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.GuideAtomBlockElement'),
	elementId: string(),
	id: string(),
	label: string(),
	title: string(),
	img: string().optional(),
	html: string(),
	credit: string(),
	role: roleTypeSchema.optional(),
});

const guVideoBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.GuVideoBlockElement'),
	elementId: string(),
	assets: array(videoAssetsSchema),
	caption: string(),
	html: string(),
	source: string(),
	role: roleTypeSchema.optional(),
	imageMedia: object({
		allImages: array(imageSchema),
	}).optional(),
	originalUrl: string().optional(),
	url: string().optional(),
});

const instagramBlockElementSchema = thirdPartyEmbeddedContentSchema.extend({
	_type: literal('model.dotcomrendering.pageElements.InstagramBlockElement'),
	elementId: string(),
	html: string(),
	url: string(),
	hasCaption: boolean(),
	role: roleTypeSchema.optional(),
});

const interactiveAtomBlockElementSchema = object({
	_type: literal(
		'model.dotcomrendering.pageElements.InteractiveAtomBlockElement',
	),
	elementId: string(),
	url: string(),
	id: string(),
	js: string().optional(),
	html: string().optional(),
	css: string().optional(),
	placeholderUrl: string().optional(),
	role: roleTypeSchema.optional(),
});

const interactiveBlockElementSchema = object({
	_type: literal(
		'model.dotcomrendering.pageElements.InteractiveBlockElement',
	),
	elementId: string(),
	url: string().optional(),
	isMandatory: boolean().optional(),
	scriptUrl: string().optional(),
	alt: string().optional(),
	role: roleTypeSchema.optional(),
	caption: string().optional(),
});

const mapBlockElementSchema = thirdPartyEmbeddedContentSchema.extend({
	_type: literal('model.dotcomrendering.pageElements.MapBlockElement'),
	elementId: string(),
	embedUrl: string(),
	originalUrl: string(),
	title: string(),
	height: number(),
	width: number(),
	caption: string().optional(),
	role: roleTypeSchema.optional(),
});

const mediaAtomBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.MediaAtomBlockElement'),
	elementId: string(),
	id: string(),
	assets: array(videoAssetsSchema),
	posterImage: array(
		object({
			url: string(),
			width: number(),
		}),
	).optional(),
	title: string().optional(),
	duration: number().optional(),
});

const newsletterSignupBlockElementSchema = object({
	_type: literal(
		'model.dotcomrendering.pageElements.NewsletterSignupBlockElement',
	),
	newsletter: newsletterSchema,
	elementId: string().optional(),
});

const interactiveContentsBlockElementSchema = object({
	_type: literal(
		'model.dotcomrendering.pageElements.InteractiveContentsBlockElement',
	),
	elementId: string(),
	subheadingLinks: array(subheadingBlockElementSchema),
	endDocumentElementId: string().optional(),
});

const profileAtomBlockElementSchema = object({
	_type: literal(
		'model.dotcomrendering.pageElements.ProfileAtomBlockElement',
	),
	elementId: string(),
	id: string(),
	label: string(),
	title: string(),
	img: string().optional(),
	html: string(),
	credit: string(),
	role: roleTypeSchema.optional(),
});

const qABlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.QABlockElement'),
	elementId: string(),
	id: string(),
	title: string(),
	img: string().optional(),
	html: string(),
	credit: string(),
	role: roleTypeSchema.optional(),
});

const richLinkBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.RichLinkBlockElement'),
	elementId: string(),
	url: string(),
	text: string(),
	prefix: string(),
	role: weightingSchema.optional(),
});

const soundcloudBlockElementSchema = thirdPartyEmbeddedContentSchema.extend({
	_type: literal('model.dotcomrendering.pageElements.SoundcloudBlockElement'),
	elementId: string(),
	html: string(),
	id: string(),
	isTrack: boolean(),
	isMandatory: boolean(),
	role: roleTypeSchema.optional(),
});

const spotifyBlockElementSchema = thirdPartyEmbeddedContentSchema.extend({
	_type: literal('model.dotcomrendering.pageElements.SpotifyBlockElement'),
	elementId: string(),
	embedUrl: string().optional(),
	title: string().optional(),
	height: number().optional(),
	width: number().optional(),
	caption: string().optional(),
	role: roleTypeSchema.optional(),
});
// ...
const calloutBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.CalloutBlockElement'),
	elementId: string(),
	id: string(),
	calloutsUrl: string(),
	activeFrom: number(),
	displayOnSensitive: boolean(),
	formId: number(),
	title: string(),
	description: string(),
	tagName: string(),
	formFields: array(campaignFieldTypeSchema),
	role: roleTypeSchema.optional(),
});

const imageBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.ImageBlockElement'),
	elementId: string(),
	media: object({
		allImages: array(imageSchema),
	}),
	data: object({
		alt: string().optional(),
		credit: string().optional(),
		caption: string().optional(),
		copyright: string().optional(),
	}),
	imageSources: array(imageSourceSchema),
	displayCredit: boolean().optional(),
	role: roleTypeSchema,
	title: string().optional(),
	starRating: number().optional(),
	isAvatar: boolean().optional(),
});

const multiImageBlockElementSchema = object({
	_type: literal('model.dotcomrendering.pageElements.MultiImageBlockElement'),
	elementId: string(),
	images: array(imageBlockElementSchema),
	caption: string().optional(),
	role: roleTypeSchema.optional(),
});

export const feElementSchema = union([
	audioAtomBlockElementSchema,
	audioBlockElementSchema,
	blockquoteBlockElementSchema,
	captionBlockElementSchema,
	calloutBlockElementSchema,
	chartAtomBlockElementSchema,
	codeBlockElementSchema,
	commentBlockElementSchema,
	contentAtomBlockElementSchema,
	disclaimerBlockElementSchema,
	dividerBlockElementSchema,
	documentBlockElementSchema,
	embedBlockElementSchema,
	explainerAtomBlockElementSchema,
	genericAtomBlockElementSchema,
	guideAtomBlockElementSchema,
	guVideoBlockElementSchema,
	highlightBlockElementSchema,
	imageBlockElementSchema,
	instagramBlockElementSchema,
	interactiveAtomBlockElementSchema,
	interactiveContentsBlockElementSchema,
	interactiveBlockElementSchema,
	itemLinkBlockElementSchema,
	mapBlockElementSchema,
	mediaAtomBlockElementSchema,
	multiImageBlockElementSchema,
	numberedTitleBlockElementSchema,
	newsletterSignupBlockElementSchema,
	profileAtomBlockElementSchema,
	pullquoteBlockElementSchema,
	qABlockElementSchema,
	quizAtomBlockElementSchema,
	richLinkBlockElementSchema,
	soundcloudBlockElementSchema,
	spotifyBlockElementSchema,
	starRatingBlockElementSchema,
	subheadingBlockElementSchema,
	tableBlockElementSchema,
	textBlockElementSchema,
	timelineBlockElementSchema,
	tweetBlockElementSchema,
	videoBlockElementSchema,
	videoFacebookBlockElementSchema,
	videoVimeoBlockElementSchema,
	videoYoutubeBlockElementSchema,
	vineBlockElementSchema,
	youtubeBlockElementSchema,
	witnessTypeBlockElementSchema,
]);